requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.6
dj-database-url==1.0.0
dj-email-url==1.0.6
Django==4.1.1
django-cache-url==3.4.2
django-ckeditor==6.5.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-js-asset==2.0.0
django-taggit==3.0.0
environs==9.5.0
marshmallow==3.18.0
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dotenv==0.21.0
sqlparse==0.4.3
whitenoise==6.2.0
Pillow==9.2.0
django-phonenumbers==1.0.1 
django-phonenumber-field==7.0.0
django-extensions==3.2.1 
pytz>=2015.7
Babel==2.10.3
django-recaptcha==3.0.0 